Move app files to [locale] folder
