+++ src/app/[locale]/dashboard/styles.module.css
