+++ src/app/[locale]/page.module.css
